+++ Fun.Blazor.Docs.Wasm/Startup.fs
@@ -11,19 +11,13 @@ open Fun.Blazor
 open Fun.Blazor.Docs.Wasm
 
 
-type PrerenderApp() =
-    inherit FunBlazorComponent()
-
-    override _.Render() = App.app
-
-
 type Program =
 
     static member ConfigureServices(services: IServiceCollection, baseAddress: string) =
         services
             .AddFunBlazorWasm()
             .AddMudServices()
-            .AddScoped<Fun.Blazor.Docs.Wasm.Demos.ScopedDemoService>()
+            .AddScoped<Demos.ScopedDemoService>()
             .AddScoped<HttpClient>(fun _ ->
                 let http = new HttpClient()
                 http.BaseAddress <- Uri baseAddress
@@ -36,13 +30,13 @@ type Program =
         let builder = WebAssemblyHostBuilder.CreateDefault(args)
 
         builder.RootComponents.RegisterForFunBlazor()
-        builder.RootComponents.RegisterCustomElementForFunBlazor<Fun.Blazor.Docs.Wasm.Demos.CustomElementDemo.DemoCounter>()
+        builder.RootComponents.RegisterCustomElementForFunBlazor<Demos.CustomElementDemo.DemoCounter>()
 
 #if DEBUG
         builder.AddFunBlazor("#app", html.hotReloadComp (app, "Fun.Blazor.Docs.Wasm.App.app"))
 #else
         if builder.RootComponents.Count = 0 then
-            builder.RootComponents.Add<PrerenderApp>("#app")
+            builder.RootComponents.Add<App>("#app")
 #endif
 
         Program.ConfigureServices(builder.Services, builder.HostEnvironment.BaseAddress)
@@ -56,4 +50,3 @@ module Startup =
 
     [<EntryPoint>]
     let main args = Program.Main args
-